+++ crates/batch-prover/src/rpc.rs
@@ -219,11 +219,16 @@ pub trait BatchProverRpc {
     ///
     /// # Arguments
     /// * `job_id` - The unique identifier of the proving job to retrieve.
+    /// * `with_proof` - Whether to include the proof in the response (default is true).
     ///
     /// # Returns
     /// An optional `JobRpcResponse` containing the job details, including commitments and proof.
     #[method(name = "getProvingJob")]
-    async fn get_proving_job(&self, job_id: Uuid) -> RpcResult<Option<JobRpcResponse>>;
+    async fn get_proving_job(
+        &self,
+        job_id: Uuid,
+        with_proof: Option<bool>,
+    ) -> RpcResult<Option<JobRpcResponse>>;
 
     /// Gets last `count` number of job ids. Returns ids in descending order, so latest job is the first index.
     ///
@@ -244,11 +249,16 @@ pub trait BatchProverRpc {
     ///
     /// # Arguments
     /// * `index` - The commitment index to retrieve the proving job for.
+    /// * `with_proof` - Whether to include the proof in the response (default is true).
     ///
     /// # Returns
     /// An optional `JobRpcResponse` containing the job details if it exists.
     #[method(name = "getProvingJobOfCommitment")]
-    async fn get_proving_job_of_commitment(&self, index: u32) -> RpcResult<Option<JobRpcResponse>>;
+    async fn get_proving_job_of_commitment(
+        &self,
+        index: u32,
+        with_proof: Option<bool>,
+    ) -> RpcResult<Option<JobRpcResponse>>;
 
     /// Gets commitment indices seen in the L1 block
     ///
@@ -557,7 +567,11 @@ where
         Ok(b64_inputs)
     }
 
-    async fn get_proving_job(&self, job_id: Uuid) -> RpcResult<Option<JobRpcResponse>> {
+    async fn get_proving_job(
+        &self,
+        job_id: Uuid,
+        with_proof: Option<bool>,
+    ) -> RpcResult<Option<JobRpcResponse>> {
         let ledger_db = &self.context.ledger_db;
 
         let Some(commitment_indices) = ledger_db
@@ -580,18 +594,22 @@ where
             });
         }
 
-        let stored_proof = ledger_db
-            .get_proof_by_job_id(job_id)
-            .map_err(internal_rpc_error)?;
+        let proof = if with_proof.unwrap_or(true) {
+            let stored_proof = ledger_db
+                .get_proof_by_job_id(job_id)
+                .map_err(internal_rpc_error)?;
 
-        let proof = match stored_proof {
-            Some(sp) => {
-                let info = ledger_db
-                    .get_proving_session_info_by_job_id(job_id)
-                    .map_err(internal_rpc_error)?;
-                Some(make_batch_proof_response(sp, info))
+            match stored_proof {
+                Some(sp) => {
+                    let info = ledger_db
+                        .get_proving_session_info_by_job_id(job_id)
+                        .map_err(internal_rpc_error)?;
+                    Some(make_batch_proof_response(sp, info))
+                }
+                None => None,
             }
-            None => None,
+        } else {
+            None
         };
 
         Ok(Some(JobRpcResponse {
@@ -622,14 +640,18 @@ where
         Ok(jobs)
     }
 
-    async fn get_proving_job_of_commitment(&self, index: u32) -> RpcResult<Option<JobRpcResponse>> {
+    async fn get_proving_job_of_commitment(
+        &self,
+        index: u32,
+        with_proof: Option<bool>,
+    ) -> RpcResult<Option<JobRpcResponse>> {
         let job_id = self
             .context
             .ledger_db
             .get_job_id_by_commitment_index(index)
             .map_err(internal_rpc_error)?;
         match job_id {
-            Some(job_id) => self.get_proving_job(job_id).await,
+            Some(job_id) => self.get_proving_job(job_id, with_proof).await,
             None => Ok(None),
         }
     }